Load recent mistakes based off of the review_statistics API
This should work for recent mistakes within 24 hrs based on the idea that if you get a mistake right again within 24 hrs, it is not longer a recent mistake. However, the extra study page now says that they always stay for 24 hours. So, I'm not sure this will always be accurate with the website. As it is, the old method is no longer accurate with the website as dates are wiped if you get them right again within 24 hours.

See: https://knowledge.wanikani.com/getting-started/extra-study/
For formula: https://community.wanikani.com/t/a-new-native-android-app/66346/232